flask/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
WTForms==3.0.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-dotenv==1.0.0
redis==5.0.0
celery==5.3.4
Pillow==10.0.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2
black==23.9.1
flake8==6.1.0
whoosh==2.7.4
jieba==0.42.1
